tqa_ui/requirements.txt
django-annoying==0.10.4
django-markdown-app==0.9.3.1
django-taggit==0.22.2
pytz==2018.3
django-hitcount==1.3.0
unidecode

# replace all .*py "from django.core.urlresolvers import reverse" to "from django.urls import reverse"